refactor: reorganize error handling and logger middleware imports
