+++ apps/api/src/middleware/logger.ts
@@ -1,5 +1,4 @@
 import { Logger } from "@repo/logger";
-import type { ErrorHandler } from "hono";
 import { createMiddleware } from "hono/factory";
 
 import type { AppEnv } from "../types";
@@ -31,18 +30,3 @@ export const loggerMiddleware = createMiddleware<AppEnv>(async (c, next) => {
     duration,
   });
 });
-
-export const onErrorHandler: ErrorHandler<AppEnv> = (err, c) => {
-  const logger = c.get("logger");
-  const requestId = c.get("requestId");
-
-  logger.error("unhandled error", {
-    requestId,
-    method: c.req.method,
-    path: c.req.path,
-    error: err.message,
-    stack: err.stack,
-  });
-
-  return c.json({ error: "Internal Server Error" }, 500);
-};
